Adicao de biblioteca de Vetores

+++ Project1/Character.h
@@ -18,7 +18,7 @@ namespace Game {
         int currentHealth = 0;
         int currentAttackPoints = 0;
         int currentDefensePoints = 0;
-        int currentSpeedPoints = 1;
+        double currentSpeedPoints = 1;
         int currentSpeedAtackPoints = 0;
 
     protected:
